TST: Reorganize plotting module.
